Add qualified Nymashock memory-card protocol profile
(cherry picked from commit 136fbcec286604a5b9597834412fd5a493b9ae19)

+++ runtime/include/sio.h
@@ -363,7 +363,7 @@ typedef struct {
     uint32_t counter_7514;   /* card chain counter byte at fire */
     uint8_t  source;         /* SioIrqSource */
     uint8_t  slot;           /* selected_slot at fire */
-    uint8_t  delay_applied;  /* the SIO_IRQ_DELAY_* value used */
+    uint16_t delay_applied;  /* cycles; source cards request 256 */
     uint8_t  pad;
 } SioIrqEntry;
